fix: compute the difference operator without exponential recursion
Each entry of the difference matrix recursed twice for each order, so a penalized fit of the difference order 24 did not finish, and the isize entries overflowed from about the order 66. The closed form with binomial coefficients gives the same values in f64.

+++ src/fit/fixed.rs
@@ -127,7 +127,7 @@ fn calculate_finite_difference_matrix(difference_order: usize, knots: &Knots) ->
 
     for i in 0..=polygon_segments - difference_order - 2 {
         for j in 0..=polygon_segments - 2 {
-            difference_matrix[(i, j)] = difference_operator(i, j, difference_order) as f64;
+            difference_matrix[(i, j)] = difference_operator(i, j, difference_order);
         }
     }
